clean up test cases
